src/requirements.txt
pi-ina219
paho-mqtt
pyyaml
